Moved files under right project
