+++ ParseIni/StateLexerCharacter.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ParseIniTests
+namespace ParseIni
 {
     public class StateLexerCharacter : IState
     {
