+++ ParseIniTests/ParseIniTests.csproj
@@ -51,12 +51,8 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="IState.cs" />
     <Compile Include="LexerTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="StateContext.cs" />
-    <Compile Include="StateLexerCharacter.cs" />
-    <Compile Include="StateLexerString.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ParseIni\ParseIni.csproj">
